requirements.txt
Flask
selenium
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv